requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
opencv-python-headless
tensorflow
keras
streamlit
tqdm

# pip install -r requirements.txt
